+++ railties/test/test_unit/reporter_test.rb
@@ -57,6 +57,38 @@ class TestUnitReporterTest < ActiveSupport::TestCase
     end
   end
 
+  test "fail fast outputs failures inline" do
+    fail_fast = Rails::TestUnitReporter.new @output, fail_fast: true
+    fail_fast.record(failed_test)
+    fail_fast.report
+
+    assert_match %r{\A\n\nboo\n\nbin/rails test .*test/test_unit/reporter_test.rb:6\n\n\z}, @output.string
+  end
+
+  test "fail fast outputs errors inline" do
+    fail_fast = Rails::TestUnitReporter.new @output, fail_fast: true
+    fail_fast.record(errored_test)
+    fail_fast.report
+
+    assert_match %r{\A\n\nArgumentError: wups\n    No backtrace\n\nbin/rails test .*test/test_unit/reporter_test.rb:6\n\n\z}, @output.string
+  end
+
+  test "fail fast outputs skipped tests inline if verbose" do
+    fail_fast = Rails::TestUnitReporter.new @output, fail_fast: true, verbose: true
+    fail_fast.record(skipped_test)
+    fail_fast.report
+
+    assert_match %r{\A\n\nskipchurches, misstemples\n\nbin/rails test .*test/test_unit/reporter_test.rb:6\n\n\z}, @output.string
+  end
+
+  test "fail fast does not output rerun snippets after run" do
+    fail_fast = Rails::TestUnitReporter.new @output, fail_fast: true
+    fail_fast.record(failed_test)
+    fail_fast.report
+
+    assert_no_match 'Failed tests:', @output.string
+  end
+
   private
   def assert_rerun_snippet_count(snippet_count)
     assert_equal snippet_count, @output.string.scan(%r{^bin/rails test }).size
@@ -72,6 +104,12 @@ class TestUnitReporterTest < ActiveSupport::TestCase
     ft
   end
 
+  def errored_test
+    et = ExampleTest.new(:woot)
+    et.failures << Minitest::UnexpectedError.new(ArgumentError.new("wups"))
+    et
+  end
+
   def passing_test
     ExampleTest.new(:woot)
   end
@@ -79,7 +117,7 @@ class TestUnitReporterTest < ActiveSupport::TestCase
   def skipped_test
     st = ExampleTest.new(:woot)
     st.failures << begin
-                     raise Minitest::Skip
+                     raise Minitest::Skip, "skipchurches, misstemples"
                    rescue Minitest::Assertion => e
                      e
                    end